parse variadic functions

+++ src/main.h
@@ -7,7 +7,7 @@
 #define TRAP() __builtin_trap()
 #endif
 #define ASSERT(x) do { if (!(x)) TRAP(); } while (0)
-#define LENGTH(x) (sizeof(x)/sizeof((x)[0]))
+#define LENGTH(x) ((isize)(sizeof(x)/sizeof((x)[0])))
 #define MIN(a, b) ((a) < (b) ? (a) : (b))
 #define MAX(a, b) ((a) > (b) ? (a) : (b))
 #define S(x) (str){(x), sizeof(x) - 1}
